update rest of Group components using Context to access groupData, setGroupData state and map over groupData for Group Name field in ExpenseForm
