+++ expense-splitter/src/components/context/SiteContext.jsx
@@ -7,8 +7,30 @@ const SiteContext = createContext(null)
 export const UseDataContext = () => useContext(SiteContext)
 
 export const DataProvider = ({children}) => {
-    const [expense, setExpense] = useState([])
+    const initalGroup = [
+        {
+          name: 'Beach Lunch',
+          description: 'Saturday lunch by the beach',
+          budget: '85',
+          id: 't0unxnqoAwLdxsjOuxc5A',
+        },
+        {
+          name: 'Bar Night',
+          description: 'Night at the bar with co-workers',
+          budget: '300',
+          id: 'NoWgbblj8apsEHZ0IVMC4',
+        },
+        {
+          name: 'Urth Cafe Brunch',
+          description: 'Brunch with the baseball team',
+          budget: '120',
+          id: 'P_PNnrbjurvmasxRtXEov',
+        },
+      ]
 
+    const [expense, setExpense] = useState([])
+    const [groupData, setGroupData] = useState(initalGroup)
+      console.log('groupData', groupData)
     const handleSetExpense = (values) => {
         setExpense(prev => [...prev, { ...values, id: nanoid(), date: new Date() }])
     }
@@ -16,7 +38,7 @@ export const DataProvider = ({children}) => {
 
 
     return (
-        <SiteContext.Provider value={{ expense, setExpense, handleSetExpense }}>
+        <SiteContext.Provider value={{ groupData, setGroupData, expense, setExpense, handleSetExpense }}>
             {children}
         </SiteContext.Provider>
     )
