+++ expense-splitter/src/components/expense/ExpenseList.jsx
@@ -1,8 +1,11 @@
 import { useState } from 'react'
+import { UseDataContext } from '../context/SiteContext'
 
-export default function ExpenseList({expense}){
-    // console.log(expense)
-    const expenseItems = expense?.map(item => (
+export default function ExpenseList(){
+   
+    const { groupData, setGroupData, expense } = UseDataContext()
+    console.log('expense', expense)
+    const expenseItems = expense.map(item => (
         <div key={item.id} className='flex flex-col bg-slate-100 rounded-md py-4 px-2 mb-1 cursor-pointer'>
 
             <div className="flex justify-between">
@@ -20,7 +23,7 @@ export default function ExpenseList({expense}){
         
     </div>
   ))
-   
+//    console.log('expense'. expense)
     
     return(
         <div className="mb-4">
