+++ expense-splitter/src/components/group/Group.jsx
@@ -2,29 +2,11 @@ import React, { useState } from 'react'
 import Button from '../ui/Button'
 import CreateGroup from './CreateGroup'
 import GroupList from './GroupList'
+import { UseDataContext } from '../context/SiteContext'
 
 export default function Group() {
-  const initalGroup = [
-    {
-      name: 'Beach Lunch',
-      description: 'Saturday lunch by the beach',
-      budget: '85',
-      id: 't0unxnqoAwLdxsjOuxc5A',
-    },
-    {
-      name: 'Bar Night',
-      description: 'Night at the bar with co-workers',
-      budget: '300',
-      id: 'NoWgbblj8apsEHZ0IVMC4',
-    },
-    {
-      name: 'Urth Cafe Brunch',
-      description: 'Brunch with the baseball team',
-      budget: '120',
-      id: 'P_PNnrbjurvmasxRtXEov',
-    },
-  ]
-  const [groupData, setGroupData] = useState(initalGroup)
+  
+  const { groupData, setGroupData } = UseDataContext()
   const [showCreateGroupForm, setShowCreateGroupForm] = useState(false)
 
   const DisplayCreateGroupForm = () => {
@@ -42,11 +24,9 @@ export default function Group() {
         </span>
       </div>
       <div>
-        <GroupList groupData={groupData} setGroupData={setGroupData} />
+        <GroupList />
         {showCreateGroupForm ? (
           <CreateGroup
-            groupData={groupData}
-            setGroupData={setGroupData}
             DisplayCreateGroupForm={DisplayCreateGroupForm}
           />
         ) : (
